moved the message to diffrent component
