+++ src/ui/pages/ServersRoom/Main.jsx
@@ -1,30 +1,13 @@
 import JoinGameRoom from "../../components/JoinGameRoom";
 import React, { useState, useEffect } from "react";
-import { CREATE_ROOM } from "../../../routes/routes_consts"
-import { useNavigate } from "react-router-dom";
+import NoRoomsMessage from "../../components/NoRoomsMessage";
 
 
 export default function Main({ rooms }) {
-  const navigate = useNavigate();
-
-  if (rooms.length === 0) {
-    return (
-      <div className="flex flex-col items-center justify-center h-full text-center p-8">
-        <p className="text-2xl sm:text-3xl font-bold mb-4 text-primary">
-          Uh-oh... no game rooms yet!
-        </p>
-        <p className="text-lg sm:text-xl mb-6">
-          Be the first to start one and invite your friends 🎉
-        </p>
-        <button
-          onClick={() => navigate(CREATE_ROOM)}
-          className="bg-primary text-white px-6 py-3 rounded-2xl shadow-lg text-sm sm:text-base hover:bg-opacity-80 hover:scale-105 hover:cursor-pointer hover:shadow-2xl transition"
-        >
-          🚀 Create a New Room
-        </button>
-      </div>
-    );
+ if (rooms.length === 0) {
+    return <NoRoomsMessage />;
   }
+  
   return (
     <main className="flex-1 overflow-y-scroll grid grid-cols-2 sm:grid-cols-3 md:grid-cols-4 gap-4 p-4 justify-items-center pb-36 sm:pb-36">
       {rooms.map((room, index) => (
